requirements.txt
# ROCK Skills Platform - Shared Dependencies
# Install: pip install -r requirements.txt

# Core dependencies
pandas>=2.0.0
numpy>=1.24.0

# AWS & Cloud
boto3>=1.28.0

# Configuration
pyyaml>=6.0

# NLP & ML (used by Project 1 & 2)
spacy>=3.5.0
# Install language model separately: python -m spacy download en_core_web_lg

# Semantic similarity (used by Project 2)
sentence-transformers>=2.2.0
scikit-learn>=1.2.0

# Graph analysis (used by Project 2 & 3)
networkx>=3.1

# LLM utilities
dicttoxml>=1.7.0  # For bedrock_client XML conversion
requests>=2.31.0

# Progress tracking
tqdm>=4.65.0

# Data validation
pydantic>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web framework (for POC app)
streamlit>=1.28.0
plotly>=5.17.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Code quality
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.23.0

# Snowflake connector (optional)
# snowflake-connector-python>=3.0.0
